refactor(rooms): extract storage, wire and authz into vti-rooms
The decomposition note asks for a named build-time or testability win
before any extraction, and says the spine is not extractable. Both apply
here, in opposite directions.

The win is a second consumer. A room host - someone hosting their own
rooms on their own infrastructure, topology T1 - stores ciphertext and
verifies presentations. Without this crate, doing that means shipping an
entire community service: member lifecycle, policy, credential issuance, a
website, an admin SPA. With it, a room host is this crate plus a dispatch
surface.

That these three modules extract cleanly is not luck, it is invariant I5
restated as a dependency graph. Because a room is authorized by
credentials the room itself issued, the code deciding a room operation
cannot need a roster, a policy engine, or a session store - so it does
not have one, and the new crate depends on vti-common and nothing else.
The compiler now enforces what was previously a convention.

The Trust-Task handlers deliberately stay in vtc-service. Dispatch is a
service's spine, and the note is explicit that a spine is not extractable;
each host writes its own thin handlers over these three layers.

The keyspace names move with the storage that uses them, and
vtc-service's registry re-exports them rather than declaring a second
copy - two hosts naming the same keyspace differently could not serve the
same room's data directory.

No behaviour change: all 18 tests moved with their modules and pass
unchanged, and vtc-service's suite is 932 where it was 950, the difference
being exactly the 18 that left.
