+++ vti-rooms/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "vti-rooms"
+description = "Data-room storage, wire types, and authorization — the parts of a room that are not a service"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+publish.workspace = true
+
+[dependencies]
+# The store abstraction and AppError. This crate's only internal dependency:
+# a room's storage and authorization need nothing from a community service,
+# which is the whole reason they are extractable.
+vti-common = { path = "../vti-common", version = "0.16" }
+
+serde = { workspace = true }
+serde_json = { workspace = true }
+
+[dev-dependencies]
+tempfile = "3"
+tokio = { workspace = true }
